修改倒计时UI

+++ VFDClock-ESP8266/WebControl.h
@@ -30,7 +30,7 @@ private:
     std::map<String, bool> userTokens; // 存储token及其登录状态
     bool isTimerRunning = false;
     int TimerMinute = 0;
-    int preTimeStamp = 0;
+    long preTimeStamp = 0;
     /**
      * main page
      */
